[FYI] ben.woodhouse, james.doverspike
#tests verified no perf regression on a less powerful platform via replay a/b tests

#ROBOMERGE-AUTHOR: jian.ru
#ROBOMERGE-SOURCE: CL 19382203 via CL 19382204 via CL 19382469 via CL 19395249 via CL 19395355
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v926-19321884)

[CL 19402948 by jian ru in ue5-main branch]

+++ Engine/Source/Runtime/Engine/Classes/Components/SkinnedMeshComponent.h
@@ -280,7 +280,7 @@ protected:
 
 protected:
 	/** The bounds radius at the point we last notified the streamer of a bounds radius change */
-	double LastStreamerUpdateBoundsRadius;
+	float LastStreamerUpdateBoundsRadius;
 
 	/** The index for the ComponentSpaceTransforms buffer we can currently write to */
 	int32 CurrentEditableComponentTransforms;
